Move IServerInterface into its own file
