+++ TGS.Interface/TGS.Interface.csproj
@@ -65,6 +65,7 @@
     <Compile Include="Enumerations.cs" />
     <Compile Include="Helpers.cs" />
     <Compile Include="InstanceMetadata.cs" />
+    <Compile Include="IServerInterface.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Components\Repository.cs" />
     <Compile Include="PullRequestInfo.cs" />
